Remove escape characters, 'I' prefix and change to single quote.
The escape characters were not needed and showed up in the code when
using the template.  The I prefix is against the coding guidelines
https://github.com/Microsoft/TypeScript/wiki/Coding-guidelines#names and
TS Lint recommends single quote for the import file name string.  Change
the Body to the Component Name so when quick prototyping you can see
which component is which.